fixed test integration
