+++ tests/embeddings_models/integration/test_integration_fasttext_embeddings_model.py
@@ -10,7 +10,7 @@ from torch.utils.data import DataLoader
 from deepparse import download_from_url
 from deepparse.embeddings_models import FastTextEmbeddingsModel
 from tests.embeddings_models.integration.tools import MockedDataTransform
-from tests.parser.integration.base_integration import AddressParserRetrainTestCase
+from tests.parser.integration.base_retrain import AddressParserRetrainTestCase
 
 
 class FastTextEmbeddingsModelIntegrationTest(AddressParserRetrainTestCase):
