+++ tests/parser/integration/test_integration_address_parser_retrain_gpu.py
@@ -10,7 +10,7 @@ import torch
 from poutyne import Callback
 
 from deepparse.parser import AddressParser
-from tests.parser.integration.base_integration import AddressParserRetrainTestCase
+from tests.parser.integration.base_retrain import AddressParserRetrainTestCase
 
 
 @skipIf(not torch.cuda.is_available(), "no gpu available")
